Update delete credentials to all multiple creds to same cluster

+++ cluster/views.py
@@ -103,16 +103,19 @@ def credential_settings(request, username):
             form = CredentialForm(request.user, initial=initial)
             usercreds = request.user.credentials.all()
             for key in request.POST:
-                if "@" in key and ":" in key and request.POST[key] == "on":
+                if "@" in key and ":" in key and '-' in key and request.POST[key] == "on":
                     username, hostname = key.split('@')
                     hostname, port = hostname.split(':')
+                    port, id_ = port.split('-')
                     try:
                         usercreds.get(
+                                    id=id_,
                                     username=username,
                                     cluster__hostname=hostname,
                                     cluster__port=int(port)).delete()
-                    except:
+                    except Exception as e:
                         pass
+            state = "Settings Successfully Saved"
         else:
             form = CredentialForm(request.user, request.POST)
             if form.is_valid():
